+++ Makefile
@@ -13,7 +13,7 @@ SRC = minirt.c $(MND)parsing/parce_rgb.c $(MND)parsing/parse_crd.c $(MND)parsing
 	$(MND)parsing/scene_obj.c $(MND)parsing/read_line.c $(MND)parsing/copy_scene_data.c $(MND)math/math1.c\
 	$(MND)parsing/err_cleanup.c $(MND)memory/arena_manage1.c $(MND)rendering/cy_loop.c $(MND)rendering/lightning.c\
 	$(MND)rendering/rotation.c $(MND)rendering/rotation_utils.c $(MND)rendering/transformation.c $(MND)rendering/move.c\
-	$(MND)rendering/move_utils.c $(MND)rendering/render_utils.c\
+	$(MND)rendering/move_utils.c $(MND)rendering/render_utils.c $(MND)rendering/hooks.c\
 
 BONUS_SRC = minirt_bonus.c  $(BNS)parsing/parce_rgb.c $(BNS)parsing/parse_crd.c $(BNS)parsing/parsing.c \
 	$(BNS)utils/ft_atof.c $(BNS)utils/ft_split.c $(BNS)utils/Rgb_cnv.c $(BNS)tmputils/cy_utils.c \
