+++ mand/src/memory/arena_create.c
@@ -6,7 +6,7 @@
 /*   By: sessarhi <sessarhi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/26 22:39:57 by sessarhi          #+#    #+#             */
-/*   Updated: 2025/01/27 15:52:47 by sessarhi         ###   ########.fr       */
+/*   Updated: 2025/02/06 20:45:43 by sessarhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,4 +45,11 @@ t_arena	*arena_create_with_params(size_t chunk_size, bool allow_resize,
 	arena->first = chunk;
 	arena->current = chunk;
 	return (arena);
+}
+
+size_t	arena_total_allocated(const t_arena	*arena)
+{
+	if (arena)
+		return (arena->total_allocated);
+	return (0);
 }
